requirements.txt
datasets==2.18.0
huggingface-hub==0.22.2
ipython==8.17.2
lightning-bolts==0.7.0
lightning-utilities==0.10.0
matplotlib==3.8.1
matplotlib-inline==0.1.6
pandas==2.1.2
pathtools==0.1.2
Pillow==10.1.0
python-dateutil==2.8.2
pytorch-lightning==1.9.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
timm==0.9.10
torchmetrics==1.3.1
tqdm==4.66.1
transformers==4.39.0
urllib3==2.0.7
wandb==0.15.12
